proyecto/systemd/requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
pycodestyle==2.7.0
pymongo==3.11.4
pyodbc==4.0.30
python-dotenv==0.17.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
